requirements.txt
numpy
GDAL
tqdm
hydra-core
omegaconf
pillow
